Release/beta.2

+++ src/index.ts
@@ -7,6 +7,29 @@ interface Coordinates {
   width: number
   height: number
 }
+
+/**
+ * Allows start/stop control of the animation
+ */
+export interface AnimationController {
+  /**
+   * The original animation parent.
+   */
+  readonly parent: Element
+  /**
+   * A function that enables future animations.
+   */
+  enable: () => void
+  /**
+   * A function that disables future animations.
+   */
+  disable: () => void
+  /**
+   * A function that returns true if the animations are currently enabled.
+   */
+  isEnabled: () => boolean
+}
+
 /**
  * A set of all the parents currently being observe. This is the only non weak
  * registry.
@@ -40,6 +63,10 @@ const options = new WeakMap<Element, AutoAnimateOptions | AutoAnimationPlugin>()
  * Debounce counters by id, used to debounce calls to update positions.
  */
 const debounces = new WeakMap<Element, NodeJS.Timeout>()
+/**
+ * All parents that are currently enabled are tracked here.
+ */
+const enabled = new WeakSet<Element>()
 /**
  * The document used to calculate transitions.
  */
@@ -125,9 +152,9 @@ function updatePos(el: Element) {
     typeof optionsOrPlugin === "function" ? 500 : optionsOrPlugin.duration
   debounces.set(
     el,
-    setTimeout(() => {
+    setTimeout(async () => {
       const currentAnimation = animations.get(el)
-      if (!currentAnimation || currentAnimation.finished) {
+      if (!currentAnimation || (await currentAnimation.finished)) {
         coords.set(el, getCoords(el))
         observePosition(el)
       }
@@ -241,7 +268,7 @@ function getElements(mutations: MutationRecord[]): Set<Element> | false {
 }
 
 /**
- *
+ * Assign the target to an element.
  * @param el - The root element
  * @param child
  */
@@ -348,6 +375,26 @@ function getOptions(el: Element): AutoAnimateOptions | AutoAnimationPlugin {
 }
 
 /**
+ * Returns the target of a given animation (generally the parent).
+ * @param el - An element to check for a target
+ * @returns
+ */
+function getTarget(el: Element): Element | undefined {
+  if (TGT in el) return (el as Element & { __aa_tgt: Element })[TGT]
+  return undefined
+}
+
+/**
+ * Checks if animations are enabled or disabled for a given element.
+ * @param el - Any element
+ * @returns
+ */
+function isEnabled(el: Element): boolean {
+  const target = getTarget(el)
+  return target ? enabled.has(target) : false
+}
+
+/**
  * Iterate over the children of a given parent.
  * @param parent - A parent element
  * @param callback - A callback
@@ -373,6 +420,7 @@ function forEach(
 function remain(el: Element) {
   const oldCoords = coords.get(el)
   const newCoords = getCoords(el)
+  if (!isEnabled(el)) return coords.set(el, newCoords)
   let animation: Animation
   if (!oldCoords) return
   const pluginOrOptions = getOptions(el)
@@ -398,7 +446,10 @@ function remain(el: Element) {
       start.height = `${heightFrom}px`
       end.height = `${heightTo}px`
     }
-    animation = el.animate([start, end], pluginOrOptions)
+    animation = el.animate([start, end], {
+      duration: pluginOrOptions.duration,
+      easing: pluginOrOptions.easing,
+    })
   } else {
     animation = new Animation(
       pluginOrOptions(el, "remain", oldCoords, newCoords)
@@ -418,6 +469,7 @@ function add(el: Element) {
   const newCoords = getCoords(el)
   coords.set(el, newCoords)
   const pluginOrOptions = getOptions(el)
+  if (!isEnabled(el)) return
   let animation: Animation
   if (typeof pluginOrOptions !== "function") {
     animation = el.animate(
@@ -452,7 +504,17 @@ function remove(el: Element) {
     next.parentNode.insertBefore(el, next)
   } else if (prev && prev.parentNode) {
     prev.parentNode.appendChild(el)
+  } else {
+    getTarget(el)?.appendChild(el)
   }
+  function cleanUp() {
+    el.remove()
+    coords.delete(el)
+    siblings.delete(el)
+    animations.delete(el)
+    intersections.get(el)?.disconnect()
+  }
+  if (!isEnabled(el)) return cleanUp()
   const [top, left, width, height] = deletePosition(el)
   const optionsOrPlugin = getOptions(el)
   const oldCoords = coords.get(el)!
@@ -487,13 +549,7 @@ function remove(el: Element) {
     animation.play()
   }
   animations.set(el, animation)
-  animation.addEventListener("finish", () => {
-    el.remove()
-    coords.delete(el)
-    siblings.delete(el)
-    animations.delete(el)
-    intersections.get(el)?.disconnect()
-  })
+  animation.addEventListener("finish", cleanUp)
 }
 
 function deletePosition(
@@ -532,6 +588,11 @@ export interface AutoAnimateOptions {
    * Default: ease-in-out
    */
   easing: "linear" | "ease-in" | "ease-out" | "ease-in-out" | string
+  /**
+   * Ignore a user’s "reduce motion" setting and enable animations. It is not
+   * recommended to use this.
+   */
+  disrespectUserMotionPreference?: boolean
 }
 
 /**
@@ -558,23 +619,38 @@ export interface AutoAnimationPlugin {
 export default function autoAnimate(
   el: HTMLElement,
   config: Partial<AutoAnimateOptions> | AutoAnimationPlugin = {}
-) {
+): AnimationController {
   if (mutations && resize) {
     const mediaQuery = window.matchMedia("(prefers-reduced-motion: reduce)")
-    if (mediaQuery.matches) return
-
-    if (getComputedStyle(el).position === "static") {
-      Object.assign(el.style, { position: "relative" })
+    const isDisabledDueToReduceMotion =
+      mediaQuery.matches &&
+      typeof config !== "function" &&
+      !config.disrespectUserMotionPreference
+    if (!isDisabledDueToReduceMotion) {
+      enabled.add(el)
+      if (getComputedStyle(el).position === "static") {
+        Object.assign(el.style, { position: "relative" })
+      }
+      forEach(el, updatePos, poll, (element) => resize?.observe(element))
+      if (typeof config === "function") {
+        options.set(el, config)
+      } else {
+        options.set(el, { duration: 250, easing: "ease-in-out", ...config })
+      }
+      mutations.observe(el, { childList: true })
+      parents.add(el)
     }
-    forEach(el, updatePos, poll, (element) => resize?.observe(element))
-    if (typeof config === "function") {
-      options.set(el, config)
-    } else {
-      options.set(el, { duration: 250, easing: "ease-in-out", ...config })
-    }
-    mutations.observe(el, { childList: true })
-    parents.add(el)
   }
+  return Object.freeze({
+    parent: el,
+    enable: () => {
+      enabled.add(el)
+    },
+    disable: () => {
+      enabled.delete(el)
+    },
+    isEnabled: () => enabled.has(el),
+  })
 }
 
 /**
@@ -583,7 +659,9 @@ export default function autoAnimate(
 export const vAutoAnimate = {
   mounted: (
     el: HTMLElement,
-    binding: { value: Partial<AutoAnimateOptions> | AutoAnimationPlugin }
+    binding: {
+      value: Partial<AutoAnimateOptions> | AutoAnimationPlugin | undefined
+    }
   ) => {
     autoAnimate(el, binding.value || {})
   },
